+++ common/store/store.js
@@ -419,7 +419,8 @@ export class StoreSyncedWithLocalStorageProvider extends React.Component {
       hasConfirmedEmail,
       hasActivatedEmail,
       disabledWarnings,
-      surveyActions
+      surveyActions,
+      userAgreementStatus
     },
     commitImmediately = true
   ) =>
@@ -437,7 +438,8 @@ export class StoreSyncedWithLocalStorageProvider extends React.Component {
             hasConfirmedEmail,
             hasActivatedEmail,
             disabledWarnings,
-            surveyActions
+            surveyActions,
+            userAgreementStatus
           }
         },
         resolve,
@@ -462,7 +464,8 @@ export class StoreSyncedWithLocalStorageProvider extends React.Component {
         name: e.company.name,
         admin: e.hasAdminRights,
         siren: e.company.siren,
-        settings: e.company.settings
+        settings: e.company.settings,
+        hasNoAdmin: e.company.hasNoActiveAdmins
       })),
       c => c.id
     );
