+++ common/utils/apiFragments.js
@@ -281,3 +281,15 @@ export const CONTROLLER_USER_FRAGMENT = gql`
     grecoId
   }
 `;
+
+export const USER_AGREEMENT = gql`
+  fragment UserAgreementData on UserAgreement {
+    hasAcceptedCgu
+    hasRejectedCgu
+    shouldAcceptCgu
+    isBlacklisted
+    answerDate
+    expiresAt
+    cguVersion
+  }
+`;
