+++ web/controller/components/modals/ControllerExportC1BAll.js
@@ -1,15 +1,8 @@
-import {
-  Button,
-  Modal,
-  ModalContent,
-  ModalFooter,
-  ModalTitle
-} from "@dataesr/react-dsfr";
+import { Button } from "@dataesr/react-dsfr";
 import { Alert } from "@mui/material";
 import Grid from "@mui/material/Grid";
 import TextField from "@mui/material/TextField";
 import Typography from "@mui/material/Typography";
-import { makeStyles } from "@mui/styles";
 import { MobileDatePicker } from "@mui/x-date-pickers";
 import { useApi } from "common/utils/api";
 import { HTTP_QUERIES } from "common/utils/apiQueries";
@@ -17,21 +10,7 @@ import { addDaysToDate, isoFormatLocalDate } from "common/utils/time";
 import React from "react";
 import { useSnackbarAlerts } from "../../../common/Snackbar";
 import SignFilesCheckbox from "../../../common/SignFiles";
-
-const useStyles = makeStyles(theme => ({
-  modalFooter: {
-    marginLeft: "auto"
-  },
-  subtitle: {
-    marginTop: theme.spacing(2),
-    marginBottom: theme.spacing(2)
-  },
-  filterGrid: {
-    paddingTop: theme.spacing(2),
-    paddingRight: theme.spacing(2),
-    flexShrink: 0
-  }
-}));
+import Modal, { modalStyles } from "../../../common/Modal";
 
 export default function ControllerExportC1BAll({
   open,
@@ -43,7 +22,7 @@ export default function ControllerExportC1BAll({
 }) {
   const alerts = useSnackbarAlerts();
   const api = useApi();
-  const classes = useStyles();
+  const classes = modalStyles();
 
   const [exportFilters, setExportFilters] = React.useState(controlFilters);
   const [sign, setSign] = React.useState(true);
@@ -52,106 +31,110 @@ export default function ControllerExportC1BAll({
   const oneYearAgo = addDaysToDate(new Date(), -365);
 
   return (
-    <Modal isOpen={open} hide={handleClose} size="lg">
-      <ModalTitle>Générez des fichiers C1B</ModalTitle>
-      <ModalContent>
-        <p>
-          Mobilic permet d'exporter les données d'activité au format C1B, qui
-          est le format utilisé pour les données des{" "}
-          <strong>cartes conducteur</strong> de{" "}
-          <strong>chronotachygraphe</strong>.
-        </p>
-        <Typography variant="h4" className={classes.subtitle}>
-          Conditions d’export
-        </Typography>
-        <p>
-          Le téléchargement produit un dossier zippé (.zip) qui contient{" "}
-          <strong>un fichier C1B pour chaque travailleur contrôlé</strong>.
-        </p>
-        <Alert severity="warning">
-          Seuls les contrôles de type "Mobilic" seront exportés au format C1B.
-        </Alert>
-        <Typography variant="h4" className={classes.subtitle}>
-          Options
-        </Typography>
-        <p>
-          Les données d'activité sont limitées à une{" "}
-          <strong>période qui ne peut pas dépasser 28 jours</strong>.
-        </p>
-        <Grid
-          spacing={2}
-          container
-          alignItems="end"
-          justifyContent="left"
-          className={classes.filterGrid}
-        >
-          <Grid item sm={6}>
-            <MobileDatePicker
-              label="Début"
-              value={exportFilters.fromDate}
-              inputFormat="d MMMM yyyy"
-              disableCloseOnSelect={false}
-              disableMaskedInput={true}
-              onChange={newFromDate => {
-                setExportFilters(prevFilters => {
-                  return {
-                    ...prevFilters,
-                    fromDate: isoFormatLocalDate(newFromDate),
-                    toDate:
-                      newFromDate > new Date(prevFilters.toDate)
-                        ? isoFormatLocalDate(newFromDate)
-                        : prevFilters.toDate
-                  };
-                });
-              }}
-              minDate={oneYearAgo}
-              maxDate={today}
-              renderInput={props => (
-                <TextField
-                  {...props}
-                  required
-                  variant="outlined"
-                  size="small"
-                />
-              )}
-            />
+    <Modal
+      open={open}
+      handleClose={handleClose}
+      title="Générez des fichiers C1B"
+      content={
+        <>
+          <p>
+            Mobilic permet d'exporter les données d'activité au format C1B, qui
+            est le format utilisé pour les données des{" "}
+            <strong>cartes conducteur</strong> de{" "}
+            <strong>chronotachygraphe</strong>.
+          </p>
+          <Typography variant="h4" className={classes.subtitle}>
+            Conditions d’export
+          </Typography>
+          <p>
+            Le téléchargement produit un dossier zippé (.zip) qui contient{" "}
+            <strong>un fichier C1B pour chaque travailleur contrôlé</strong>.
+          </p>
+          <Alert severity="warning">
+            Seuls les contrôles de type "Mobilic" seront exportés au format C1B.
+          </Alert>
+          <Typography variant="h4" className={classes.subtitle}>
+            Options
+          </Typography>
+          <p>
+            Les données d'activité sont limitées à une{" "}
+            <strong>période qui ne peut pas dépasser 28 jours</strong>.
+          </p>
+          <Grid
+            spacing={2}
+            container
+            alignItems="end"
+            justifyContent="left"
+            className={classes.filterGrid}
+          >
+            <Grid item sm={6}>
+              <MobileDatePicker
+                label="Début"
+                value={exportFilters.fromDate}
+                inputFormat="d MMMM yyyy"
+                disableCloseOnSelect={false}
+                disableMaskedInput={true}
+                onChange={newFromDate => {
+                  setExportFilters(prevFilters => {
+                    return {
+                      ...prevFilters,
+                      fromDate: isoFormatLocalDate(newFromDate),
+                      toDate:
+                        newFromDate > new Date(prevFilters.toDate)
+                          ? isoFormatLocalDate(newFromDate)
+                          : prevFilters.toDate
+                    };
+                  });
+                }}
+                minDate={oneYearAgo}
+                maxDate={today}
+                renderInput={props => (
+                  <TextField
+                    {...props}
+                    required
+                    variant="outlined"
+                    size="small"
+                  />
+                )}
+              />
+            </Grid>
+            <Grid item sm={6}>
+              <MobileDatePicker
+                label="Fin"
+                value={exportFilters.toDate}
+                inputFormat="d MMMM yyyy"
+                disableCloseOnSelect={false}
+                disableMaskedInput={true}
+                onChange={newToDate => {
+                  setExportFilters(prevFilters => {
+                    return {
+                      ...prevFilters,
+                      toDate: isoFormatLocalDate(newToDate),
+                      fromDate:
+                        newToDate < new Date(prevFilters.fromDate)
+                          ? isoFormatLocalDate(newToDate)
+                          : prevFilters.fromDate
+                    };
+                  });
+                }}
+                maxDate={today}
+                renderInput={props => (
+                  <TextField
+                    {...props}
+                    required
+                    variant="outlined"
+                    size="small"
+                  />
+                )}
+              />
+            </Grid>
+            <Grid item xs={12}>
+              <SignFilesCheckbox sign={sign} setSign={setSign} />
+            </Grid>
           </Grid>
-          <Grid item sm={6}>
-            <MobileDatePicker
-              label="Fin"
-              value={exportFilters.toDate}
-              inputFormat="d MMMM yyyy"
-              disableCloseOnSelect={false}
-              disableMaskedInput={true}
-              onChange={newToDate => {
-                setExportFilters(prevFilters => {
-                  return {
-                    ...prevFilters,
-                    toDate: isoFormatLocalDate(newToDate),
-                    fromDate:
-                      newToDate < new Date(prevFilters.fromDate)
-                        ? isoFormatLocalDate(newToDate)
-                        : prevFilters.fromDate
-                  };
-                });
-              }}
-              maxDate={today}
-              renderInput={props => (
-                <TextField
-                  {...props}
-                  required
-                  variant="outlined"
-                  size="small"
-                />
-              )}
-            />
-          </Grid>
-          <Grid item xs={12}>
-            <SignFilesCheckbox sign={sign} setSign={setSign} />
-          </Grid>
-        </Grid>
-      </ModalContent>
-      <ModalFooter>
+        </>
+      }
+      actions={
         <Button
           title="téléchargement"
           className={classes.modalFooter}
@@ -170,7 +153,7 @@ export default function ControllerExportC1BAll({
         >
           Générer
         </Button>
-      </ModalFooter>
-    </Modal>
+      }
+    />
   );
 }
