+++ web/controller/components/modals/ControllerHelp.js
@@ -1,14 +1,15 @@
 import React from "react";
-import { Modal, ModalTitle, ModalContent } from "@dataesr/react-dsfr";
 import { HelpController } from "../help/ModalHelpController";
+import Modal from "../../../common/Modal";
 
 export default function ControllerHelp({ open, handleClose }) {
   return (
-    <Modal isOpen={open} hide={handleClose} size="lg">
-      <ModalTitle>Besoin d'aide ?</ModalTitle>
-      <ModalContent>
-        <HelpController />
-      </ModalContent>
-    </Modal>
+    <Modal
+      open={open}
+      handleClose={handleClose}
+      title="Besoin d'aide ?"
+      content={<HelpController />}
+      size="sm"
+    />
   );
 }
