+++ web/pwa/modals/NewMissionModal.js
@@ -1,11 +1,11 @@
 import React from "react";
-import { FunnelModal } from "./FunnelModal";
 import NewMissionForm from "../../common/NewMissionForm";
 import { useStoreSyncedWithLocalStorage } from "common/store/store";
 import { DISMISSABLE_WARNINGS } from "../../admin/utils/dismissableWarnings";
 import { useModals } from "common/utils/modals";
 import { useSnackbarAlerts } from "../../common/Snackbar";
 import { setCurrentLocation } from "common/utils/location";
+import { FunnelModal } from "../components/FunnelModal";
 
 export default function NewMissionModal({
   open,
@@ -16,7 +16,8 @@ export default function NewMissionModal({
   disableCurrentPosition = false,
   disableKilometerReading = false,
   withDay = false,
-  withEndLocation = false
+  withEndLocation = false,
+  onSelectNoAdminCompany = null
 }) {
   const [currentPosition, setCurrentPosition] = React.useState(null);
 
@@ -58,6 +59,10 @@ export default function NewMissionModal({
         withEndLocation={withEndLocation}
         askCurrentPosition={askCurrentPosition}
         disableGeolocation={disableCurrentPosition}
+        onSelectNoAdminCompany={() => {
+          handleClose();
+          onSelectNoAdminCompany();
+        }}
       />
     </FunnelModal>
   );
