+++ web/pwa/screens/BeforeWork.js
@@ -46,26 +46,11 @@ const useStyles = makeStyles(theme => ({
     backgroundSize: "cover",
     position: "relative"
   },
-  container: {
-    position: "relative",
-    justifyContent: "flex-start",
-    flexGrow: 1,
-    flexShrink: 0,
-    overflowY: "hidden"
-  },
-  heroImage: {
-    width: "100%",
-    height: "auto",
-    color: theme.palette.background.default
-  },
   accountButton: {
     alignSelf: "flex-end",
     position: "absolute",
     top: "0",
     right: "0"
-  },
-  welcomeText: {
-    marginBottom: theme.spacing(2)
   },
   missionsToValidateList: {
     backgroundColor: theme.palette.background.paper,
@@ -136,6 +121,11 @@ export function BeforeWork({ beginNewMission, openHistory, missions }) {
   const userId = store.userId();
   const userInfo = store.userInfo();
 
+  const areAllCompaniesWithoutAdmins = React.useMemo(
+    () => companies.every(c => !!c.hasNoAdmin),
+    [companies]
+  );
+
   function handleFirstActivitySelection(teamMates, missionInfos) {
     const team = teamMates ? [userId, ...teamMates.map(cw => cw.id)] : [userId];
     modals.open("firstActivity", {
@@ -171,27 +161,32 @@ export function BeforeWork({ beginNewMission, openHistory, missions }) {
   }
 
   const onEnterNewMissionFunnel = () => {
-    modals.open("newMission", {
-      companies,
-      companyAddresses: store.getEntity("knownAddresses"),
-      handleContinue: missionInfos => {
-        const company = companies.find(c => c.id === missionInfos.company.id);
-        if (company.settings && company.settings.allowTeamMode) {
-          modals.open("teamOrSoloChoice", {
-            handleContinue: isTeamMode => {
-              if (isTeamMode) {
-                modals.open("teamSelection", {
-                  mission: null,
-                  companyId: missionInfos.company.id,
-                  handleContinue: teamMates =>
-                    handleFirstActivitySelection(teamMates, missionInfos)
-                });
-              } else handleFirstActivitySelection(null, missionInfos);
-            }
-          });
-        } else handleFirstActivitySelection(null, missionInfos);
-      }
-    });
+    if (areAllCompaniesWithoutAdmins) {
+      modals.open("blockedTime", {});
+    } else {
+      modals.open("newMission", {
+        companies,
+        companyAddresses: store.getEntity("knownAddresses"),
+        handleContinue: missionInfos => {
+          const company = companies.find(c => c.id === missionInfos.company.id);
+          if (company.settings && company.settings.allowTeamMode) {
+            modals.open("teamOrSoloChoice", {
+              handleContinue: isTeamMode => {
+                if (isTeamMode) {
+                  modals.open("teamSelection", {
+                    mission: null,
+                    companyId: missionInfos.company.id,
+                    handleContinue: teamMates =>
+                      handleFirstActivitySelection(teamMates, missionInfos)
+                  });
+                } else handleFirstActivitySelection(null, missionInfos);
+              }
+            });
+          } else handleFirstActivitySelection(null, missionInfos);
+        },
+        onSelectNoAdminCompany: () => modals.open("blockedTime", {})
+      });
+    }
   };
 
   const onEnterNewHolidayFunnel = () => openHolidaysModal();
@@ -230,7 +225,7 @@ export function BeforeWork({ beginNewMission, openHistory, missions }) {
       <PlaceHolder>
         <img alt="mobilic-logo-text" src={LogoWithText} width={150} />
       </PlaceHolder>
-      <Box key={2} mt={2} className="cta-container" mb={2}>
+      <Box mt={2} className="cta-container" mb={2}>
         <Typography className={classes.promiseText}>
           Le suivi de votre temps de travail
         </Typography>
@@ -271,7 +266,7 @@ export function BeforeWork({ beginNewMission, openHistory, missions }) {
         </LoadingButton>
       </Box>
       {nonValidatedMissions.length > 0 && (
-        <Box key={3} className={classes.missionsToValidateList}>
+        <Box className={classes.missionsToValidateList}>
           <List
             subheader={
               <ListSubheader
