+++ web/root.js
@@ -67,6 +67,8 @@ import { loadControllerUserData } from "./controller/utils/loadControllerUserDat
 import { LocalizationProvider, frFR } from "@mui/x-date-pickers";
 import { shouldUpdatePassword } from "common/utils/updatePassword";
 import UpdatePasswordModal from "./pwa/components/UpdatePassword";
+import AcceptCguModal from "./pwa/modals/AcceptCguModal";
+import RejectedCguModal from "./pwa/modals/RejectedCguModals";
 
 const matomo = createInstance({
   urlBase: "https://stats.beta.gouv.fr",
@@ -292,6 +294,17 @@ function _Root() {
     return () => {};
   }, [controllerId]);
 
+  const [seeAgainCgu, setSeeAgainCgu] = React.useState(false);
+
+  const shouldSeeCguModal = React.useMemo(
+    () => userInfo?.userAgreementStatus?.shouldAcceptCgu || seeAgainCgu,
+    [userInfo?.userAgreementStatus?.shouldAcceptCgu, seeAgainCgu]
+  );
+  const shouldSeeHasRejectedCguModal = React.useMemo(
+    () => !seeAgainCgu && userInfo?.userAgreementStatus?.hasRejectedCgu,
+    [userInfo?.userAgreementStatus?.hasRejectedCgu, seeAgainCgu]
+  );
+
   const routes = getAccessibleRoutes({ userInfo, companies, controllerInfo });
 
   return (
@@ -303,6 +316,19 @@ function _Root() {
       {process.env.REACT_APP_CRISP_AUTOLOAD !== "1" &&
         process.env.REACT_APP_CRISP_WEBSITE_ID &&
         !controllerId && <LiveChat />}
+      {store.userId() && shouldSeeCguModal && (
+        <AcceptCguModal
+          onAccept={() => setSeeAgainCgu(false)}
+          onReject={() => setSeeAgainCgu(false)}
+        />
+      )}
+      {store.userId() && shouldSeeHasRejectedCguModal && (
+        <RejectedCguModal
+          expiryDate={userInfo?.userAgreementStatus?.expiresAt}
+          onRevert={() => setSeeAgainCgu(true)}
+          userId={userId}
+        />
+      )}
       {store.userId() && shouldUpdatePassword() && <UpdatePasswordModal />}
       <React.Suspense fallback={<CircularProgress color="primary" />}>
         <Switch color="secondary">
